Safe cast to bring pointer arithmetic in line with a much smaller   result set.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@170464 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/unix/inet_pton.c
@@ -116,7 +116,7 @@ inet_pton4(const char *src, unsigned char *dst)
 	const char *pch;
 
 	if ((pch = strchr(digits, ch)) != NULL) {
-	    unsigned int new = *tp * 10 + (pch - digits);
+	    unsigned int new = *tp * 10 + (unsigned int)(pch - digits);
 
 	    if (new > 255)
 		return (0);
